+++ vite.config.ts
@@ -2,11 +2,8 @@ import { fileURLToPath, URL } from 'node:url';
 import { defineConfig } from 'vite';
 import dts from 'vite-plugin-dts';
 
-function fileName(format: string): string {
-  if (format === 'es') {
-    return 'index.mjs';
-  }
-  return 'index.js';
+function fileName(_format: string, entryName: string): string {
+  return `${entryName}.mjs`;
 }
 
 export default defineConfig({
@@ -14,6 +11,7 @@ export default defineConfig({
     dts({
       entryRoot: 'src',
       outDir: 'dist',
+      exclude: ['**/*.spec.ts', '**/test-helpers.ts'],
       insertTypesEntry: true,
     }),
   ],
@@ -24,8 +22,11 @@ export default defineConfig({
   },
   build: {
     lib: {
-      entry: fileURLToPath(new URL('./src/index.ts', import.meta.url)),
-      formats: ['es', 'cjs'],
+      entry: {
+        index: fileURLToPath(new URL('./src/index.ts', import.meta.url)),
+        vite: fileURLToPath(new URL('./src/vite.ts', import.meta.url)),
+      },
+      formats: ['es'],
       fileName,
     },
     sourcemap: true,